backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pandas
werkzeug==2.3.7
gunicorn==21.2.0
python-magic==0.4.27